fix(speedcurve): self-review fixes
