+++ test/e2e/basic.test.ts
@@ -637,190 +637,3 @@ describe.skip('social-embeds', () => {
     expect(hasProxiedImages).toBe(true)
   })
 })
-
-describe('speedcurve', { timeout: 15000 }, () => {
-  // Block the real CDN so window.LUX spy set up in fixtures is never overwritten
-  async function scPage(path: string) {
-    return createPage(path, {
-      setup: (p: any) => p.route('**/cdn.speedcurve.com/**', (r: any) => r.abort()),
-    })
-  }
-
-  // Helper: read captured LUX calls from the page's spy
-  async function getLuxCalls(page: any) {
-    return page.evaluate(() => (window as any)._luxCalls as Array<{ method: string, args: unknown[] }>)
-  }
-
-  // Helper: wait for a specific LUX method to appear in _luxCalls
-  async function waitForLuxCall(page: any, method: string, timeoutMs = 5000) {
-    await page.waitForFunction(
-      (m: string) => ((window as any)._luxCalls as any[])?.some((c: any) => c.method === m),
-      method,
-      { timeout: timeoutMs },
-    )
-  }
-
-  it('primer: injects LUX into <head> with correct snippetVersion', async () => {
-    const { page } = await scPage('/tpc/speedcurve')
-
-    const snippetVersion = await page.evaluate(
-      () => (window as any).LUX?.snippetVersion,
-    )
-    expect(snippetVersion).toBeTruthy()
-    expect(String(snippetVersion)).toMatch(/^\d+\.\d+/)
-  })
-
-  // eslint-disable-next-line test/prefer-lowercase-title
-  it('SPA auto: calls startSoftNavigation when navigating to a new route', async () => {
-    const { page } = await scPage('/tpc/speedcurve')
-
-    // Navigate via NuxtLink (SPA navigation, no full reload)
-    await page.click('#nav-destination')
-    await page.waitForSelector('#page')
-
-    await waitForLuxCall(page, 'startSoftNavigation')
-    const calls = await getLuxCalls(page)
-    const softNavCalls = calls.filter(c => c.method === 'startSoftNavigation')
-    expect(softNavCalls).toHaveLength(1)
-  })
-
-  // eslint-disable-next-line test/prefer-lowercase-title
-  it('SPA auto: calls markLoadTime after page:finish + paint', async () => {
-    const { page } = await scPage('/tpc/speedcurve')
-
-    await page.click('#nav-destination')
-    await page.waitForSelector('#page')
-
-    // markLoadTime fires after page:finish + two rAFs — give it time
-    await waitForLuxCall(page, 'markLoadTime', 8000)
-    const calls = await getLuxCalls(page)
-    const loadTimeCalls = calls.filter(c => c.method === 'markLoadTime')
-    expect(loadTimeCalls.length).toBeGreaterThanOrEqual(1)
-  })
-
-  // eslint-disable-next-line test/prefer-lowercase-title
-  it('SPA auto: timing order is startSoftNavigation before markLoadTime', async () => {
-    const { page } = await scPage('/tpc/speedcurve')
-    // Wait for the initial page:finish markLoadTime (fired via rAF) to land,
-    // then clear so only the next navigation's calls are in the array.
-    await waitForLuxCall(page, 'markLoadTime', 3000)
-    await page.evaluate(() => { (window as any)._luxCalls = [] })
-
-    await page.click('#nav-destination')
-    await page.waitForSelector('#page')
-    await waitForLuxCall(page, 'markLoadTime', 8000)
-
-    const calls = await getLuxCalls(page)
-    const softNavIdx = calls.findIndex(c => c.method === 'startSoftNavigation')
-    const markLoadIdx = calls.findIndex(c => c.method === 'markLoadTime')
-    expect(softNavIdx).toBeGreaterThanOrEqual(0)
-    expect(markLoadIdx).toBeGreaterThan(softNavIdx)
-  })
-
-  // eslint-disable-next-line test/prefer-lowercase-title
-  it('SPA auto default label: uses Nuxt-generated route name as label', async () => {
-    const { page } = await scPage('/tpc/speedcurve')
-
-    await page.click('#nav-destination')
-    await page.waitForSelector('#page')
-
-    const label = await page.evaluate(() => (window as any).LUX?.label)
-    // Nuxt generates a route name from the file path — e.g. 'tpc-speedcurve-destination'.
-    // String(to.name ?? to.path) returns this name. Path fallback (to.name === undefined)
-    // is unreachable via Nuxt file-based routing — it is covered by unit test instead.
-    expect(typeof label).toBe('string')
-    expect(label).toMatch(/speedcurve.*destination/)
-  })
-
-  // eslint-disable-next-line test/prefer-lowercase-title
-  it('SPA auto: back navigation triggers second startSoftNavigation', {
-    timeout: 20000,
-  }, async () => {
-    const { page } = await scPage('/tpc/speedcurve')
-
-    await page.click('#nav-destination')
-    await page.waitForSelector('#page')
-    await waitForLuxCall(page, 'markLoadTime', 8000)
-
-    await page.click('#nav-back')
-    await page.waitForSelector('#nav-destination')
-
-    await waitForLuxCall(page, 'startSoftNavigation')
-    const calls = await getLuxCalls(page)
-    const softNavCalls = calls.filter(c => c.method === 'startSoftNavigation')
-    expect(softNavCalls.length).toBeGreaterThanOrEqual(2)
-  })
-
-  it('canceled navigation: seals phantom beacon with luxNavFailed tag', async () => {
-    const { page } = await scPage('/tpc/speedcurve')
-
-    // A guard in the fixture blocks navigation to /blocked and returns false
-    await page.click('#nav-blocked')
-
-    // The afterEach(failure) handler fires synchronously after the guard blocks
-    await page.waitForFunction(
-      () => ((window as any)._luxCalls as any[])?.some((c: any) => c.method === 'addData'),
-      { timeout: 5000 },
-    )
-
-    const calls = await getLuxCalls(page)
-    const addDataCall = calls.find(c => c.method === 'addData')
-    expect(addDataCall?.args).toEqual(['luxNavFailed', '1'])
-
-    const markLoadCall = calls.find(c => c.method === 'markLoadTime')
-    expect(markLoadCall).toBeDefined()
-
-    // We must still be on the index page (navigation was blocked)
-    expect(page.url()).toContain('/tpc/speedcurve')
-    expect(page.url()).not.toContain('/blocked')
-  })
-
-  // eslint-disable-next-line test/prefer-lowercase-title
-  it('SPA off: no startSoftNavigation call when spaMode is false', async () => {
-    const { page } = await scPage('/tpc/speedcurve-no-spa')
-
-    await page.click('#nav-destination')
-    await page.waitForSelector('#page')
-    await page.waitForTimeout(1000) // give hooks time to fire if they mistakenly exist
-
-    const calls = await getLuxCalls(page)
-    const softNavCalls = calls.filter(c => c.method === 'startSoftNavigation')
-    expect(softNavCalls).toHaveLength(0)
-  })
-
-  it('custom labelFor: applies the user-provided label function', async () => {
-    const { page } = await scPage('/tpc/speedcurve-custom-label')
-
-    await page.click('#nav-destination')
-    await page.waitForSelector('#page')
-
-    await page.waitForTimeout(500)
-    const label = await page.evaluate(() => (window as any).LUX?.label)
-    expect(label).toBe('custom:/tpc/speedcurve-custom-label/destination')
-  })
-
-  it('labelFor false: leaves window.LUX.label unchanged after navigation', async () => {
-    const { page } = await scPage('/tpc/speedcurve-label-off')
-
-    await page.click('#nav-destination')
-    await page.waitForSelector('#page')
-
-    await page.waitForTimeout(500)
-    const label = await page.evaluate(() => (window as any).LUX?.label)
-    // The fixture initialises label to 'original-label'; with labelFor:false the composable
-    // must never overwrite it, regardless of which route was navigated to.
-    expect(label).toBe('original-label')
-  })
-
-  it('autoTrackSpaNavigations false: no startSoftNavigation calls on navigation', { timeout: 20000 }, async () => {
-    const { page } = await scPage('/tpc/speedcurve-manual')
-
-    await page.click('#nav-destination')
-    await page.waitForSelector('#page')
-    await page.waitForTimeout(1000) // give hooks time to fire if they mistakenly exist
-
-    const calls = await getLuxCalls(page)
-    const softNavCalls = calls.filter(c => c.method === 'startSoftNavigation')
-    expect(softNavCalls).toHaveLength(0)
-  })
-})
