+++ test/fixtures/speedcurve/app.vue
@@ -1,0 +1,3 @@
+<template>
+  <NuxtPage />
+</template>
