+++ test/fixtures/speedcurve/pages/tpc/speedcurve-custom-label/destination.vue
