+++ test/fixtures/speedcurve/pages/tpc/speedcurve-custom-label/index.vue
@@ -1,8 +1,8 @@
 <script setup lang="ts">
 if (import.meta.client) {
-  window._luxCalls = []
+  ;(window as any)._luxCalls = []
   const record = (method: string) => (...args: unknown[]) => {
-    window._luxCalls!.push({ method, args })
+    ;(window as any)._luxCalls.push({ method, args })
   }
   window.LUX = {
     snippetVersion: '2.0.0',
@@ -16,13 +16,15 @@ if (import.meta.client) {
 
 const { status } = useScriptSpeedCurve({
   id: '123456789',
-  spaMode: false,
+  spaMode: true,
+  autoTrackSpaNavigations: true,
+  labelFor: to => `custom:${to.path}`,
 })
 </script>
 
 <template>
   <div>
     <div id="status">{{ status }}</div>
-    <NuxtLink id="nav-destination" to="/tpc/speedcurve-no-spa/destination">destination</NuxtLink>
+    <NuxtLink id="nav-destination" to="/tpc/speedcurve-custom-label/destination">destination</NuxtLink>
   </div>
 </template>
