+++ test/fixtures/speedcurve/pages/tpc/speedcurve-label-off/destination.vue
