+++ test/fixtures/speedcurve/pages/tpc/speedcurve-manual/destination.vue
