+++ test/fixtures/speedcurve/pages/tpc/speedcurve-no-spa/destination.vue
