+++ test/fixtures/speedcurve/pages/tpc/speedcurve/index.vue
@@ -1,10 +1,10 @@
 <script setup lang="ts">
 // Install LUX spy BEFORE composable runs.
 // The primer does `window.LUX = window.LUX || {}` — pre-existing object is preserved.
-if (import.meta.client && !window._luxCalls) {
-  window._luxCalls = []
+if (import.meta.client && !(window as any)._luxCalls) {
+  ;(window as any)._luxCalls = []
   const record = (method: string) => (...args: unknown[]) => {
-    window._luxCalls!.push({ method, args })
+    ;(window as any)._luxCalls.push({ method, args })
   }
   window.LUX = {
     snippetVersion: '2.0.0',
